--- src/node/desktop/test/unit/core/system.test.ts
@@ -1,46 +1,0 @@
-/*
- * system.test.ts
- *
- * Copyright (C) 2021 by RStudio, PBC
- *
- * Unless you have received this program directly from RStudio pursuant
- * to the terms of a commercial license agreement with RStudio, then
- * this program is licensed to you under the terms of version 3 of the
- * GNU Affero General Public License. This program is distributed WITHOUT
- * ANY EXPRESS OR IMPLIED WARRANTY, INCLUDING THOSE OF NON-INFRINGEMENT,
- * MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE. Please refer to the
- * AGPL (http://www.gnu.org/licenses/agpl-3.0.txt) for more details.
- *
- */
-
-import { describe } from 'mocha';
-import { assert } from 'chai';
-
-import fs from 'fs';
-import os from 'os';
-
-import { FilePath } from '../../../src/core/file-path';
-import { initHook, initializeLog, userHomePath, username } from '../../../src/core/system';
-import * as log from '../../../src/core/log';
-
-describe('System', () => {
-  describe('User info', () => {
-    it('getUserHomePath returns a valid path', () => {
-      assert.isTrue(fs.existsSync(userHomePath().getAbsolutePath()));
-    });
-    it('username returns a non-empty string', () => {
-      assert.isNotEmpty(username());
-    });
-  });
-  describe('Assorted helper functions', () => {
-    it('initHook does something', () => {
-      initHook();
-      // TODO: once this does something, test it!
-    });
-    it('initializeLog succeeds', () => {
-      const err = initializeLog('rdesktop', log.LogLevel.WARN, new FilePath(os.tmpdir()));
-      assert.isFalse(!!err);
-    });
-  });
-});
-
